charlas/requirements.txt
PyYaml
Jinja2
